# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Style": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "style_contract": {
            "title": "Style Contract",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "style_contract",
          "created_at"
        ],
        "title": "Style",
        "type": "object"
      },
      "StyleInput": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "style_contract": {
            "default": "",
            "title": "Style Contract",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "StyleInput",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LoreCraft3000",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/styles": {
      "get": {
        "operationId": "list_styles_api_v1_styles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Style"
                  },
                  "title": "Response List Styles Api V1 Styles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Styles",
        "tags": [
          "api-v1",
          "api-v1-styles"
        ]
      },
      "post": {
        "operationId": "create_style_api_v1_styles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Style"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Style",
        "tags": [
          "api-v1",
          "api-v1-styles"
        ]
      }
    },
...
    "/api/v1/styles/{style_id}": {
      "delete": {
        "description": "Archive (soft-delete) a style. Existing scenes keep referencing it.",
        "operationId": "archive_style_api_v1_styles__style_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "style_id",
            "required": true,
            "schema": {
              "title": "Style Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Style",
        "tags": [
          "api-v1",
          "api-v1-styles"
        ]
      },
      "get": {
        "operationId": "get_style_api_v1_styles__style_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "style_id",
            "required": true,
            "schema": {
              "title": "Style Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Style"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Style",
        "tags": [
          "api-v1",
          "api-v1-styles"
        ]
      },
      "put": {
        "operationId": "update_style_api_v1_styles__style_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "style_id",
            "required": true,
            "schema": {
              "title": "Style Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Style"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Style",
        "tags": [
          "api-v1",
          "api-v1-styles"
        ]
      }
    },
    "/api/v1/styles/{style_id}/restore": {
      "post": {
        "operationId": "restore_style_api_v1_styles__style_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "style_id",
            "required": true,
            "schema": {
              "title": "Style Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Style"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Style",
        "tags": [
          "api-v1",
          "api-v1-styles"
        ]
      }
    },
...
  }
}